+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Versions with only mechanical changes will be omitted from the following list.
 * Fix `DurationRound` is not TZ aware (#495)
 * Implement `DurationRound` for `NaiveDateTime`
 * Add `DateTime::from_local()` to construct from given local date and time (#572)
+* Fix the problem that format will panic when given invalid format string (#614)
 
 ## 0.4.19